% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flw_palettes.R
\name{generate_palette_flw}
\alias{generate_palette_flw}
\title{Generate a color palette based on Fallingwater or Taliesin schemes}
\usage{
generate_palette_flw(
  palette_name = c("fallingwater", "taliesin"),
  selected_color
)
}
\arguments{
\item{palette_name}{Character, choose between "fallingwater" or "taliesin"}

\item{selected_color}{Hex color to center the palette around (run \code{\link{show_all_colors_flw}} to see options)}
}
\value{
A vector of neighboring colors; A ggplot object is attached as an attribute called "plot"
}
\description{
Generate a color palette based on Fallingwater or Taliesin schemes
}
\examples{
palette <- generate_palette_flw("fallingwater", "#ad3a3f")
palette # this will show the palette visually and display the hex codes in the R console
}
