% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flw_palettes.R
\name{show_all_colors_flw}
\alias{show_all_colors_flw}
\title{Display a color palette based on Fallingwater or Taliesin schemes}
\usage{
show_all_colors_flw(palette_name = c("fallingwater", "taliesin"))
}
\arguments{
\item{palette_name}{Character, choose between "fallingwater" or "taliesin"}
}
\value{
None; A ggplot of the color palette will be displayed
}
\description{
Display a color palette based on Fallingwater or Taliesin schemes
}
\examples{
show_all_colors_flw(palette_name = 'fallingwater')
}
